# swagger.yaml
openapi: 3.0.0
info:
  title: safe-courier
  description: Safe Courier is a courier service that helps users deliver parcels to different destinations.
  contact:
    name: Atuhaire Collins
    email: collinsbenda360@gmail.com
  version: '1.0'
servers:
  - url: http://localhost:5000/api/v1
    variables: {}
paths:
...
  /parcels:
    get:
      tags:
        - Parcels
      summary: get all parcels
      operationId: getallparcels
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
        - Parcels
      summary: create parcel
      operationId: createparcel
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createparcelrequest'
            example:
              receiver:
                name: collins
                email: collins2@gmail.com
                phoneNumber: 0770567432
                address: kampala
              _id: 61a09d1832f0da1db182254f
              parcelType: safeCaurier
              description: delivery app
              weight: 10
              weightUnit: kgs
              sender: 61a0a791c32916c03bdb0d95
              pickupLocation: usa
              destination: kampala
              presentLocation: kampala
              status: processing
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    createparcelrequest:
      title: createparcelrequest
      required:
        - receiver
        - _id
        - parcelType
        - description
        - weight
        - weightUnit
        - sender
        - pickupLocation
        - destination
        - presentLocation
        - status
      type: object
      properties:
        receiver:
          $ref: '#/components/schemas/Receiver'
        _id:
          type: string
        parcelType:
          type: string
        description:
          type: string
        weight:
          type: integer
          format: int32
        weightUnit:
          type: string
        sender:
          type: string
        pickupLocation:
          type: string
        destination:
          type: string
        presentLocation:
          type: string
        status:
          type: string
      example:
        receiver:
          name: collins
          email: collins2@gmail.com
          phoneNumber: 0770567432
          address: kampala
        _id: 61a09d1832f0da1db182254f
        parcelType: safeCaurier
        description: delivery app
        weight: 10
        weightUnit: kgs
        sender: 61a0a791c32916c03bdb0d95
        pickupLocation: usa
        destination: kampala
        presentLocation: kampala
        status: processing
    Receiver:
      title: Receiver
      required:
        - name
        - email
        - phoneNumber
        - address
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
      example:
        name: collins
        email: collins2@gmail.com
        phoneNumber: 0770567432
        address: kampala
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...